feat: State Hoisting     - Hoist state from screens     - Introduce routes for holding state and action
